cohorts_api/requirements.txt
psycopg2-binary==2.8.6
pandas==0.25.3
Flask==1.1.1
numpy==1.17.4